// ui/src/generated/CollabCommand.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PresenceState } from "./PresenceState";

export type CollabCommand = { "type": "Join", server: string, session: string, token: string | null, name: string, } | { "type": "Leave" } | { "type": "SetPresence", presence: PresenceState, };